refactor: Splits up the main file
* Logic separated into 3 files for better organization and logic separation moving forward.
